Fix for #25: iconRect.y calculated from iconSize.width not iconSize.height.
This commit fixes an issue where setting a non-square icon size for a
Source List could cause the icons to display incorrectly. This was
because the min-y value used when calculating icon rects was being based
on the width set for the Source List's icon size, rather than the height.