Implement +[PXSourceList methodNameMappingsForProtocol:]
This method gets all the method names from the given protocol and creates a dictionary
mapping from these methods to those which will be used by the Source List delegate and
data source. In particular, it replaces occurrences of "outlineView" with "sourceList" in
a case-insensitive way and matches the case of the "o" and the "v" in the replacement.