Purge NSOutlineView's -respondsToSelector: cache when changing the delegate and data source.
When the source list delegate and data source changes we want NSOutlineView to remove any
cached data it has on what methods its delegate and data source respond to. Since PXSourceList
is always the outline view's delegate and data source, we need to call -setDelegate: and -setDataSource:
on super first with nil then back to the PXSourceList instance again.